+++ src/components/buttons/TabLink/TabLink.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import { useMatch } from 'react-router'
+import { useMatch, useResolvedPath } from 'react-router'
 
 import { Button, ButtonProps } from '../Button/Button'
 
@@ -13,7 +13,9 @@ export const TabLink: FC<TabLinkProps> = props => {
   const theme = useTheme()
   const { children, ...otherProps } = props
   const { to } = otherProps
-  const match = useMatch({ path: to })
+
+  const resolved = useResolvedPath(to)
+  const match = useMatch({ path: resolved.pathname, end: true })
 
   return (
     <Button {...otherProps} isRaw css={[defaultButtonStyles(theme), match ? activeButtonStyles(theme) : null]}>